+++ Particle.py
@@ -359,3 +359,108 @@ class Particle:
     
     def __len__(self):
         return len(self.px)
+
+
+
+class Electron(Particle):
+    from typing import Optional
+    import numpy as np
+    def __init__(self, x:            Optional[np.ndarray]           = None, \
+                       y:            Optional[np.ndarray]           = None, \
+                       tof:          Optional[np.ndarray]           = None, \
+                       px:           Optional[np.ndarray]           = None, \
+                       py:           Optional[np.ndarray]           = None, \
+                       pz:           Optional[np.ndarray]           = None, \
+                       p:            Optional[np.ndarray]           = None, \
+                       energy:       Optional[np.ndarray]           = None, \
+                       spectrometer: Optional[Spectrometer]         = None, \
+                       calcSettings: Optional[CalcSettings]         = None, \
+                       dtype:  np.typing.DTypeLike                  = np.double, \
+                       ctype:  np.typing.DTypeLike                  = np.cdouble, \
+                ) -> None:
+        super().__init__(x=x, y=y, tof=tof, m=1, q=-1, px=px, py=py, pz=py, \
+                         p=p, energy=energy, spectrometer=spectrometer,     \
+                         calcSettings=calcSettings, isIonSide=False,        \
+                         dtype=dtype, ctype=ctype)
+        
+
+
+class Ion(Particle):
+    from typing import Optional
+    import numpy as np
+    def __init__(self, x:            Optional[np.ndarray]           = None, \
+                       y:            Optional[np.ndarray]           = None, \
+                       tof:          Optional[np.ndarray]           = None, \
+                       m:            Optional[np.ndarray|int|float] = None, \
+                       q:            Optional[np.ndarray|int|float] = None, \
+                       tofMean:      Optional[float]                = None, \
+                       px:           Optional[np.ndarray]           = None, \
+                       py:           Optional[np.ndarray]           = None, \
+                       pz:           Optional[np.ndarray]           = None, \
+                       p:            Optional[np.ndarray]           = None, \
+                       energy:       Optional[np.ndarray]           = None, \
+                       spectrometer: Optional[Spectrometer]         = None, \
+                       calcSettings: Optional[CalcSettings]         = None, \
+                       isIonSide:    Optional[bool]                 = True, \
+                       dtype:  np.typing.DTypeLike                  = np.double, \
+                       ctype:  np.typing.DTypeLike                  = np.cdouble, \
+                ) -> None:
+        m *= CONSTANTS.U_SI_TO_KG_SI
+        m /= CONSTANTS.ME_SI_TO_AU
+        super().__init__(x=x, y=y, tof=tof, m=m, q=q, px=px, py=py, pz=py, \
+                         p=p, energy=energy, spectrometer=spectrometer,    \
+                         calcSettings=calcSettings, isIonSide=isIonSide,   \
+                         dtype=dtype, ctype=ctype, tofMean=tofMean)
+           
+                         
+
+class ParticleList:
+    from typing import Optional
+    def __init__(self, particles: Optional[list[Particle]] = None):
+        self.particles = list() if particles is None else particles
+    
+    def __len__(self) -> int:
+        return len(self.particles)
+    
+    def __iter__(self) -> ParticleList:
+        self.iterIndex = 0
+        return self
+    
+    def __next__(self) -> Particle:
+        if self.iterIndex < len(self):
+            val = self.particles[self.iterIndex]
+            self.iterIndex += 1
+            return val
+        else:
+            raise StopIteration
+    
+    def __getitem__(self, index) -> Particle:
+        return self.particles[index]
+    
+    
+    def __add__(self, other: ParticleList|Particle) -> ParticleList:
+        """
+        Returns a shallow copy of the particle list, with other appended.
+        """
+        copy = self.particle.copy()
+        if   isinstance(other, ParticleList):
+            copy.extend(other)
+            return copy
+        elif isinstance(other, Particle):
+            copy.append(other)
+            return copy
+        else:
+            raise NotImplementedError
+    
+    def __iadd__(self, other: ParticleList|Particle) -> ParticleList:
+        """
+        Appends other to the list of particles.
+        """
+        if   isinstance(other, ParticleList):
+            self.particles.extend(other)
+            return self
+        elif isinstance(other, Particle):
+            self.particles.append(other)
+            return self
+        else:
+            raise NotImplementedError
